refactor: migrate repeater to typescript
